requirements/requirements.txt
django-environ==0.4.5
django-extensions==2.2.5
Werkzeug==0.16.0
Django==2.2.*
psycopg2==2.8.*
djangorestframework==3.10.*
coreapi==2.3.*
Markdown==3.1.*
Pygments==2.4.*
django-filter==2.1.*
drf-yasg
drf-yasg[validation]
pillow==7.2.*
pandas==0.25.*
xlrd==1.2.*